requirements.txt
fastapi
uvicorn
solders
python-telegram-bot
solana
aiohttp
